code clean up

+++ WXKit/WXKit/UIImage+WXKit.m
@@ -12,9 +12,9 @@
 
 - (UIImage *)imageScaleToSize:(CGSize)size
 {
-    float screenScale = [UIScreen mainScreen].scale;
+    CGFloat screenScale = [UIScreen mainScreen].scale;
     size = CGSizeMake(size.width * screenScale, size.height * screenScale);
-    float scale = size.width / self.size.width;
+    CGFloat scale = size.width / self.size.width;
     if (self.size.height * scale > size.height || self.size.width * scale > size.width) scale = size.height / self.size.height;
     UIImage *scaledImage = [UIImage imageWithCGImage:self.CGImage scale: 1 / scale orientation:self.imageOrientation];
     return scaledImage;
